Change imports after moving variable explorer editor widgets.
