+++ spyder/plugins/variableexplorer/widgets/namespacebrowser.py
@@ -30,7 +30,7 @@ from spyder.utils.misc import fix_reference_name
 from spyder.utils.programs import is_module_installed
 from spyder.utils.qthelpers import (add_actions, create_action,
                                     create_toolbutton)
-from spyder.plugins.variableexplorer.widgets.collectionseditor import (
+from spyder.plugins.variableexplorer.widgets.viewers.collections import (
     RemoteCollectionsEditorTableView)
 from spyder.plugins.variableexplorer.widgets.importwizard import ImportWizard
 from spyder.plugins.variableexplorer.widgets.utils import REMOTE_SETTINGS
