+++ spyder/plugins/variableexplorer/widgets/viewers/object.py
@@ -58,14 +58,14 @@ def create_dialog(obj, obj_name):
     oedit to show eMZed related data)
     """
     # Local import
-    from spyder.plugins.variableexplorer.widgets.texteditor import TextEditor
+    from spyder.plugins.variableexplorer.widgets.viewers.text import TextEditor
     from spyder.plugins.variableexplorer.widgets.utils import (ndarray, FakeObject,
                                                Image, is_known_type, DataFrame,
                                                Series)
-    from spyder.plugins.variableexplorer.widgets.collectionseditor import CollectionsEditor
-    from spyder.plugins.variableexplorer.widgets.arrayeditor import ArrayEditor
+    from spyder.plugins.variableexplorer.widgets.viewers.collections import CollectionsEditor
+    from spyder.plugins.variableexplorer.widgets.viewers.array import ArrayEditor
     if DataFrame is not FakeObject:
-        from spyder.plugins.variableexplorer.widgets.dataframeeditor import DataFrameEditor
+        from spyder.plugins.variableexplorer.widgets.viewers.dataframe import DataFrameEditor
 
     conv_func = lambda data: data
     readonly = not is_known_type(obj)
